clean redundant code & rename class
