+++ Slither_Project/lib/widgets/MainUI.dart
@@ -3,8 +3,6 @@ import 'package:flutter/material.dart';
 
 import '../Answer/Answer.dart';
 import '../Front/EnterScene.dart';
-import '../Scene/GameSceneSquareProvider.dart';
-import '../Scene/GameSceneSquare_Instance.dart';
 import '../Scene/GameSceneSquare.dart';
 import '../ThemeColor.dart';
 import '../User/UserInfo.dart';
@@ -410,7 +408,7 @@ class MainUI {
         Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) => GameSceneSquareProvider(isContinue: isContinue, loadKey: key),
+              builder: (context) => GameSceneSquare(isContinue: isContinue, loadKey: key),
             )
         );
         break;
